requirements.txt
pandas
openpyxl
python-dotenv
sib-api-v3-sdk
